\name{plasma}
\alias{plasma}
\title{plasma}
\usage{plasma}
\description{
The data are collected from 273 female patients.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{y}{plasma beta-carotene}
    \item{calories}{calories in KJ}
    \item{dietary}{dietary beta-carotene}
  }
}
\source{
Jiang, Y., Wang, Y-G., Fu, L., & Wang, X. (2018). Robust Estimation Using Modified Huber's Functions With New Tails. Technometrics, in press, see <doi:10.1080/00401706.2018.1470037>.
}

\keyword{datasets}
